Move metadata assertions closer to its subject
