+++ flogger/flogger-api/src/test/kotlin/io/spine/logging/flogger/LogContextSpec.kt
@@ -31,13 +31,13 @@ import io.spine.logging.flogger.LogContext.Key
 import io.spine.logging.flogger.LogContext.specializeLogSiteKeyFromMetadata
 import io.spine.logging.flogger.context.Tags
 import io.spine.logging.flogger.given.iterate
-import com.google.common.flogger.testing.shouldContainInOrder
-import com.google.common.flogger.testing.shouldContain
+import io.spine.logging.flogger.backend.given.shouldContainInOrder
+import io.spine.logging.flogger.backend.given.shouldContain
 import io.spine.logging.flogger.given.shouldHaveArguments
 import io.spine.logging.flogger.given.shouldHaveMessage
-import com.google.common.flogger.testing.shouldHaveSize
-import com.google.common.flogger.testing.shouldNotContain
-import com.google.common.flogger.testing.shouldUniquelyContain
+import io.spine.logging.flogger.backend.given.shouldHaveSize
+import io.spine.logging.flogger.backend.given.shouldNotContain
+import io.spine.logging.flogger.backend.given.shouldUniquelyContain
 import com.google.common.flogger.testing.FakeLogSite
 import com.google.common.flogger.testing.FakeLoggerBackend
 import com.google.common.flogger.testing.FakeMetadata
